# handlers/branches.go
package handlers

import (
	"encoding/json"
	"fmt"
	"net/http"
	"strconv"

	"github.com/aiman-zaki/go_dz_http/models"
	"github.com/aiman-zaki/go_dz_http/services"
	"github.com/aiman-zaki/go_dz_http/wrappers"
	"github.com/go-chi/chi"
	"github.com/go-pg/pg/v9"
)

type BranchResources struct{}

func (rs BranchResources) Routes() chi.Router {
	r := chi.NewRouter()
	r.Route("/", func(r chi.Router) {
		// swagger:route GET /branches Branches getBranches
		//
		// Get all Branches.
		//
		//    Consumes;
		//     - application/json
		//    Produces:
		//     - application/json
		//    Schemes: http, https
		//
		//    Security:
		//      Bearer:
		//     Responses:
		//       200:branches
		//
		r.Get("/", rs.Read)
		// swagger:route GET /branches/{id} Branches getBranchById
		//
		// Get a Branch by Id.
		//
		//    Consumes;
		//     - application/json
		//    Produces:
		//     - application/json
		//    Schemes: http, https
		//
		//    Security:
		//      Bearer:
		//     Responses:
		//       200:branches
		//
		r.Get("/{id}", rs.ReadByID)
		// swagger:route POST /branches Branches createBranch
		//
		// Create a Branch.
		//
		//    Consumes;
		//     - application/json
		//    Produces:
		//     - application/json
		//    Schemes: http, https
		//
		//    Security:
		//      Bearer:
		//     Responses:
		//       200:branch
		//
		r.Post("/", rs.Create)
		// swagger:route PUT /branches/{id} Branches updateBranchById
		//
		// Update a Branches.
		//
		//    Consumes;
		//     - application/json
		//    Produces:
		//     - application/json
		//    Schemes: http, https
		//
		//    Security:
		//      Bearer:
		//     Responses:
		//       200:branch
		//
		r.Put("/{id}", rs.Update)
		// swagger:route DELETE /branches/{id} Branches deleteBranchById
		//
		// Delete a Branch by Id.
		//
		//    Consumes;
		//     - application/json
		//    Produces:
		//     - application/json
		//    Schemes: http, https
		//
		//    Security:
		//      Bearer:
		//     Responses:
		//       200:branch
		//
		r.Delete("/{id}", rs.Delete)
	})

	return r
}

func (rs BranchResources) CountBranchExist(db *pg.DB, id int64, m models.Branch) int {

	count, err := db.Model(&m).Where("id = ?", id).SelectAndCount()
	if err != nil {
		fmt.Println(err)
	}
	return count
}

func (rs BranchResources) Create(w http.ResponseWriter, r *http.Request) {
	var bw models.BranchWrapper
	db := pg.Connect(services.PgOptions())
	defer db.Close()
	wrappers.JSONDecodeWrapper(w, r, &bw.Single)
	bw.Create()
	json.NewEncoder(w).Encode(bw.Single)

}

func (rs BranchResources) Update(w http.ResponseWriter, r *http.Request) {
	var bw models.BranchWrapper
	wrappers.JSONDecodeWrapper(w, r, &bw.Single)
	id := chi.URLParam(r, "id")
	parsedID, parseErr := strconv.ParseInt(id, 0, 64)
	if parseErr != nil {
		http.Error(w, `{"message":"invalid id format"}`, 400)
		return
	}
	bw.Single.ID = parsedID
	bw.Update()
	json.NewEncoder(w).Encode(bw.Update)
}

func (rs BranchResources) Read(w http.ResponseWriter, r *http.Request) {
	var bw models.BranchWrapper
	err := bw.Read()
	if err != nil {
		http.Error(w, err.Error(), 400)
		return
	}
	json.NewEncoder(w).Encode(bw.Array)
}

func (rs BranchResources) ReadByID(w http.ResponseWriter, r *http.Request) {
	var bw models.BranchWrapper
	id := chi.URLParam(r, "id")
	parseID, _ := strconv.ParseInt(id, 0, 64)
	bw.Single.ID = parseID
	bw.ReadById()
	json.NewEncoder(w).Encode(bw.Single)
}

func (rs BranchResources) Delete(w http.ResponseWriter, r *http.Request) {
	var bw models.BranchWrapper
	id := chi.URLParam(r, "id")
	parseID, _ := strconv.ParseInt(id, 0, 64)
	bw.Single.ID = parseID
	err := bw.Delete()
	if err != nil {
		http.Error(w, err.Error(), 400)
	}
	json.NewEncoder(w).Encode(bw.Single)

}