# swagger_ui/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "DZ API.",
    "version": "0.0.1"
  },
  "host": "localhost:8181",
  "basePath": "/api/",
  "paths": {
    "/auth/login": {
      "post": {
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Authentication"
        ],
        "summary": "Login",
        "operationId": "authLogin",
        "parameters": [
          {
            "name": "Auth",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/validCredential"
          },
          "401": {
            "$ref": "#/responses/invalidCredential"
          },
          "404": {
            "$ref": "#/responses/errorNotFound"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Authentication"
        ],
        "summary": "Register",
        "operationId": "authRegister",
        "parameters": [
          {
            "name": "Auth",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/validCredential"
          },
          "422": {
            "$ref": "#/responses/dataAlreadyExisted"
          }
        }
      }
    },
    "/branches": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Branches"
        ],
        "summary": "Get all Branches.",
        "operationId": "getBranches",
        "responses": {
          "200": {
            "$ref": "#/responses/branches"
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Branches"
        ],
        "summary": "Create a Branch.",
        "operationId": "createBranch",
        "parameters": [
          {
            "name": "Branch",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Branch"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/branch"
          }
        }
      }
    },
    "/branches/{id}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Branches"
        ],
        "summary": "Get a Branch by Id.",
        "operationId": "getBranchById",
        "responses": {
          "200": {
            "$ref": "#/responses/branches"
          }
        }
      },
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Branches"
        ],
        "summary": "Update a Branches.",
        "operationId": "updateBranchById",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/branch"
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Branches"
        ],
        "summary": "Delete a Branch by Id.",
        "operationId": "deleteBranchById",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/branch"
          }
        }
      }
    },
...
    "/products": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Products"
        ],
        "summary": "Get all Products.",
        "operationId": "getProducts",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "CurrentPage",
            "name": "currentPage",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "PerPage",
            "name": "perPage",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/products"
          },
          "401": {
            "$ref": "#/responses/notAuthorized"
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Products"
        ],
        "summary": "Create a Product.",
        "operationId": "createProduct",
        "parameters": [
          {
            "name": "Product",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/product"
          },
          "401": {
            "$ref": "#/responses/notAuthorized"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Products"
        ],
        "summary": "GET a Product by ID.",
        "operationId": "getProductById",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/product"
          },
          "401": {
            "$ref": "#/responses/notAuthorized"
          }
        }
      },
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Products"
        ],
        "summary": "Update a Product.",
        "operationId": "updateProduct",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "CurrentPage",
            "name": "currentPage",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "PerPage",
            "name": "perPage",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/product"
          },
          "401": {
            "$ref": "#/responses/notAuthorized"
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Products"
        ],
        "summary": "DELETE a Product by ID.",
        "operationId": "deleteProductById",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/product"
          },
          "401": {
            "$ref": "#/responses/notAuthorized"
          }
        }
      }
    },
...
    "/role": {
      "get": {
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Roles"
        ],
        "summary": "Get All Roles",
        "operationId": "getRoles",
        "responses": {
          "200": {
            "$ref": "#/responses/roles"
          }
        }
      },
      "post": {
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Roles"
        ],
        "summary": "Create New Role",
        "operationId": "createRole",
        "parameters": [
          {
            "name": "Role",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Role"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/role"
          },
          "422": {
            "$ref": "#/responses/dataAlreadyExisted"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Auth": {
      "description": "Auth represents the Authentication Model for this application",
      "type": "object",
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "access_token": {
          "description": "accesstoken (jwt) expired (1 hours)",
          "type": "string",
          "x-go-name": "AcessToken",
          "readOnly": true
        },
        "email": {
          "description": "email for auth",
          "type": "string",
          "x-go-name": "Email"
        },
        "id": {
          "description": "id for user",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID",
          "readOnly": true
        },
        "password": {
          "description": "password for auth",
          "type": "string",
          "x-go-name": "Password"
        },
        "refresh_token": {
          "description": "refreshtoken (jwt) expired (15 days)",
          "type": "string",
          "x-go-name": "RefreshToken",
          "readOnly": true
        }
      },
      "x-go-package": "github.com/aiman-zaki/go_dz_http/models"
    },
    "Branch": {
      "description": "Branch is a model bro",
      "type": "object",
      "properties": {
        "address": {
          "type": "string",
          "x-go-name": "Address"
        },
        "branch": {
          "type": "string",
          "x-go-name": "Branch"
        },
        "date_created": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "DateCreated"
        },
        "date_updated": {
          "description": "the dateUpdated for the product",
          "type": "string",
          "format": "date-time",
          "x-go-name": "DateUpdated"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID",
          "readOnly": true
        }
      },
      "x-go-package": "github.com/aiman-zaki/go_dz_http/models"
    },
...
    "Product": {
      "description": "Product represents the product for this application",
      "type": "object",
      "properties": {
        "cost_price": {
          "description": "the dateCreated for the product",
          "type": "number",
          "format": "float",
          "x-go-name": "CostPrice"
        },
        "date_created": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "DateCreated"
        },
        "date_updated": {
          "description": "the dateUpdated for the product",
          "type": "string",
          "format": "date-time",
          "x-go-name": "DateUpdated"
        },
        "id": {
          "description": "the id for the product",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "product": {
          "description": "the name for the product",
          "type": "string",
          "x-go-name": "Product"
        },
        "sale_price": {
          "type": "number",
          "format": "float",
          "x-go-name": "SalePrice"
        }
      },
      "x-go-package": "github.com/aiman-zaki/go_dz_http/models"
    },
    "Role": {
      "description": "Role : model for the role existes in the system",
      "type": "object",
      "properties": {
        "id": {
          "description": "id for role",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID",
          "readOnly": true
        },
        "key": {
          "description": "the role",
          "type": "string",
          "x-go-name": "Key"
        },
        "text": {
          "type": "string",
          "x-go-name": "Text"
        }
      },
      "x-go-package": "github.com/aiman-zaki/go_dz_http/models"
    },
...
  },
  "responses": {
    "branch": {
      "description": "BranchResponse : List a branch",
      "schema": {
        "type": "object",
        "properties": {
          "branch": {
            "$ref": "#/definitions/Branch"
          },
          "message": {
            "description": "the success message",
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
    "branches": {
      "description": "BranchesResponse : List all branches",
      "schema": {
        "type": "object",
        "properties": {
          "branches": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Branch"
            },
            "x-go-name": "Branch"
          },
          "message": {
            "description": "the success message",
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
    "dataAlreadyExisted": {
      "description": "DataAlreadyExisted :",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
    "errorNotFound": {
      "description": "ErrorNotFound :",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
    "invalidCredential": {
      "description": "InvalidCredential :",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "description": "the error message",
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
    "notAuthorized": {
      "description": "NotAuthorized :",
      "schema": {
        "type": "object",
        "properties": {
          "Message": {
            "type": "string"
          }
        }
      }
    },
...
    "product": {
      "description": "ProductResponse : List a product",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "description": "the success message",
            "type": "string",
            "x-go-name": "Message"
          },
          "product": {
            "$ref": "#/definitions/Product"
          }
        }
      }
    },
    "products": {
      "description": "ProductsResponse : List all products",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "description": "the success message",
            "type": "string",
            "x-go-name": "Message"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Product"
            },
            "x-go-name": "Product"
          }
        }
      }
    },
    "role": {
      "description": "RoleResponse :",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "description": "the success message",
            "type": "string",
            "x-go-name": "Message"
          },
          "role": {
            "$ref": "#/definitions/Role"
          }
        }
      }
    },
    "roles": {
      "description": "RolesResponse :",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "description": "the success message",
            "type": "string",
            "x-go-name": "Message"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Role"
            },
            "x-go-name": "Role"
          }
        }
      }
    },
...
    "validCredential": {
      "description": "ValidCredential :",
      "schema": {
        "type": "object",
        "properties": {
          "auth": {
            "$ref": "#/definitions/Auth"
          },
          "message": {
            "description": "the success message",
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "flow": "implicit"
    }
  },
  "security": [
    {
      "Bearer": [
        "[]"
      ]
    }
  ]
}